+++ integration_tests/suite/test_agents.py
@@ -1,14 +1,25 @@
 # Copyright 2019-2020 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from hamcrest import all_of, assert_that, calling, has_properties, is_, matches_regexp
+import time
+
+from hamcrest import (
+    all_of,
+    assert_that,
+    calling,
+    has_properties,
+    is_,
+    matches_regexp,
+)
 
 from wazo_agentd_client.error import (
     AgentdClientError,
+    NOT_LOGGED,
     NO_SUCH_AGENT,
     NO_SUCH_LINE,
     UNAUTHORIZED,
 )
+from xivo_test_helpers import until
 from xivo_test_helpers.hamcrest.raises import raises
 
 from .helpers.base import BaseIntegrationTest, UNKNOWN_UUID, UNKNOWN_ID
@@ -90,7 +101,9 @@ class TestAgents(BaseIntegrationTest):
 
     @fixtures.user_line_extension(exten='1001', context='default', name_line='abcdef')
     @fixtures.agent(number='1234')
-    def test_login_logoff_user_agent_on_specific_line(self, user_line_extension, agent):
+    def test_login_pause_unpause_logoff_user_agent_on_specific_line(
+        self, user_line_extension, agent
+    ):
         self.create_user_token(user_line_extension['user_uuid'])
 
         with associations.user_agent(
@@ -99,13 +112,14 @@ class TestAgents(BaseIntegrationTest):
             # login
             self.agentd.agents.login_user_agent(user_line_extension['line_id'])
 
-            status = self.agentd.agents.get_agent_status(agent['id'])
+            status = self.agentd.agents.get_user_agent_status()
             assert_that(
                 status,
                 has_properties(
                     {
                         'id': agent['id'],
                         'logged': True,
+                        'paused': False,
                         'context': 'default',
                         'extension': '1001',
                         'number': '1234',
@@ -114,16 +128,40 @@ class TestAgents(BaseIntegrationTest):
                 ),
             )
 
+            # pause
+            self.agentd.agents.pause_user_agent()
+            # NOTE(fblackburn): agentd may be still not connected to the bus to receive messages
+            # A best solution should be to create a /status and wait on it
+            time.sleep(8)
+            self.bus.send_queue_member_pause('1234', paused=True)
+
+            def test_on_msg_received():
+                status = self.agentd.agents.get_user_agent_status()
+                assert_that(status, has_properties(paused=True))
+
+            until.assert_(test_on_msg_received, tries=10)
+
+            # unpause
+            self.agentd.agents.unpause_user_agent()
+            self.bus.send_queue_member_pause('1234', paused=False)
+
+            def test_on_msg_received():
+                status = self.agentd.agents.get_user_agent_status()
+                assert_that(status, has_properties(paused=False))
+
+            until.assert_(test_on_msg_received, tries=10)
+
             # logoff
             self.agentd.agents.logoff_user_agent()
 
-            status = self.agentd.agents.get_agent_status(agent['id'])
+            status = self.agentd.agents.get_user_agent_status()
             assert_that(
                 status,
                 has_properties(
                     {
                         'id': agent['id'],
                         'logged': False,
+                        'paused': False,
                         'context': None,
                         'extension': None,
                         'number': '1234',
@@ -164,8 +202,16 @@ class TestAgents(BaseIntegrationTest):
             )
 
     @fixtures.user_line_extension(exten='1001', context='default', name_line='abcdef')
-    def test_login_logoff_user_agent_without_agent(self, user_line_extension):
+    def test_get_login_pause_unpause_logoff_user_agent_without_agent(
+        self, user_line_extension
+    ):
         self.create_user_token(user_line_extension['user_uuid'])
+
+        # get
+        assert_that(
+            calling(self.agentd.agents.get_user_agent_status),
+            raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
+        )
 
         # login
         assert_that(
@@ -174,6 +220,19 @@ class TestAgents(BaseIntegrationTest):
             ),
             raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
         )
+
+        # pause
+        assert_that(
+            calling(self.agentd.agents.pause_user_agent),
+            raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
+        )
+
+        # unpause
+        assert_that(
+            calling(self.agentd.agents.unpause_user_agent),
+            raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
+        )
+
         # logoff
         assert_that(
             calling(self.agentd.agents.logoff_user_agent),
@@ -181,8 +240,16 @@ class TestAgents(BaseIntegrationTest):
         )
 
     @fixtures.user_line_extension(exten='1001', context='default', name_line='abcdef')
-    def test_login_logoff_user_agent_without_user(self, user_line_extension):
+    def test_login_pause_unpause_logoff_user_agent_without_user(
+        self, user_line_extension
+    ):
         self.create_user_token(UNKNOWN_UUID)
+
+        # get
+        assert_that(
+            calling(self.agentd.agents.get_user_agent_status),
+            raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
+        )
 
         # login
         assert_that(
@@ -192,8 +259,59 @@ class TestAgents(BaseIntegrationTest):
             raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
         )
 
+        # pause
+        assert_that(
+            calling(self.agentd.agents.pause_user_agent),
+            raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
+        )
+
+        # unpause
+        assert_that(
+            calling(self.agentd.agents.unpause_user_agent),
+            raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
+        )
+
         # logoff
         assert_that(
             calling(self.agentd.agents.logoff_user_agent),
             raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
+        )
+
+    @fixtures.agent(id=42, number='1234')
+    @fixtures.user_line_extension(
+        agentid=42, exten='1001', context='default', name_line='ab'
+    )
+    def test_pause_unpause_logoff_user_agent_not_logged(
+        self, agent, user_line_extension
+    ):
+        self.create_user_token(user_line_extension['user_uuid'])
+
+        # pause
+        assert_that(
+            calling(self.agentd.agents.pause_agent_by_number).with_args('1234'),
+            raises(AgentdClientError, has_properties(error=NOT_LOGGED)),
+        )
+        assert_that(
+            calling(self.agentd.agents.pause_user_agent),
+            raises(AgentdClientError, has_properties(error=NOT_LOGGED)),
+        )
+
+        # unpause
+        assert_that(
+            calling(self.agentd.agents.unpause_agent_by_number).with_args('1234'),
+            raises(AgentdClientError, has_properties(error=NOT_LOGGED)),
+        )
+        assert_that(
+            calling(self.agentd.agents.unpause_user_agent),
+            raises(AgentdClientError, has_properties(error=NOT_LOGGED)),
+        )
+
+        # logoff
+        assert_that(
+            calling(self.agentd.agents.logoff_agent_by_number).with_args('1234'),
+            raises(AgentdClientError, has_properties(error=NOT_LOGGED)),
+        )
+        assert_that(
+            calling(self.agentd.agents.logoff_user_agent),
+            raises(AgentdClientError, has_properties(error=NOT_LOGGED)),
         )